requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
jinja2>=3.0.0
python-multipart>=0.0.5
aiofiles>=23.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
reportlab>=4.0.0
